Change DuckDbVectorInfo* argument in converters to "in DbDuckVectorInfo"
This was the original intent but the code with the read-only reference
argument did not compile earlier.  It might have been because the generic
type T involved was allowed to be a "ref struct" (and function pointers,
used to invoke the conversion function, do not allowed the "scoped"
modifier).  Now "ref structs" are ruled out so the C# compiler complains
no more.