Introduce ConverterCreationContext
This begins to re-organize the code for the converters so that they
can be cached outside of any vector.

Basically, factory methods for VariantElementConverter that took
DuckDbVectorInfo now take ConverterCreationContext instead.

There is no actual vector-independent caching yet, so the eventual
contents of ConverterCreationContext has been commented out,
and it is just a wrapper for the native DuckDB vector object (handle),
temporarily.

The list converter is (the only) one converter that is not vector-independent
right now, so we cannot change over to the new scheme right away.