Fix wrong argument type in 2 methods called by function pointer
There was no crash because the ABI of the argument (pass by reference) happen
to be the same and the argument is ignored, but obviously the code needs to be
fixed.  It got introduced as part of recent work on making converters work
across multiple vectors (for the same column).