Combine DuckDbStatement.Parameter.SetNativeValue with its helper method
... in the parent class.  The helper method was defined in the parent
class because in the old parameters API, that was the place to put it.

There is only one caller to the helper method so its code can be merged
into its caller.  This move also fixes a subtle bug: if the instance of
DuckDbStatement.Parameter is default-initialized (always an
incorrect client usage), the native resource will not have been freed
before the method exited with NullReferenceException.